Echo path in og:url
Facebook will use the Open Graph metadata for canonical URL in link
previews. This is probably intended behavior for Open Graph. Setting the
canonical URL to the root host will ruin the link preview and make it
not match the actual path, so set the canonical URl to the exact same
URL used to request the page.